requirements.txt
django==2.2.4
djangorestframework==3.10.2
scikit-learn==0.21.3
Keras==2.2.4
tensorflow==1.14.0
psycopg2-binary==2.8.4
django-crispy-forms==1.7.2